4.0 Branch Merge (#63)
* make Invoke-IBWAPI emit non-terminating errors rather than terminating exceptions
* started testing batch mode for New/Remove-IBObject
* added batch mode for Set-IBObject
* added readable fields schema cache to speed up -ReturnAllFields queries
* added batch mode for Get-IBObject with pipelined object refs
* tweaked verbose/debug logging
* fixed Get-IBObject tests to account for schema cache
* bump the min PS version to 5.1 and 7.0 and add System.Web to manifest req assemblies
* Renamed Filters param to Filter (with Filters alias for back compat) and supports using a hashtable instead of a string array that will be automatically URL encoded
* Invoke-IBWAPI will now handle converting Body params to JSON and UTF-8 encoding as long as ContentType not specified so calling functions don't have to do it in advance. Body content is now always included in Verbose output.
* updated instdev.ps1 for ps 5.1 minimum
* fixed tests to not overwrite local config
* cleaned up desktop edition check
* refactored more error handling
* refined error categories
* removed support for old 1.x profiles
* added support for env variable based profile
* added SecretManagement support, still needs vault specific tests
* added some vault specific tests for Export-IBConfig
* added some vault specific tests for Import-IBConfig
* added new param set to Invoke-IBWAPI that takes individual host/version/query rather than a raw Uri and migrated internal stuff to use it
* converted comment-based help to platyPS markdown and MAML
* added the rest of a basic docs site
* doc tweaks
* added stateless mode guide
* added guide for SecretManagement
* Filter param will take any IDictionary object instead of just [hashtable]
* added preliminary 4.0 changelog
* added -BatchGroupSize which defaults to 1000
* fixed doc refs that still referred to -Filters
* renamed -ReturnAllFields to -ReturnAll
* renamed -ReturnBaseFields to -ReturnBase
* renamed -ReturnFields to -ReturnField
* tweaked some param ordering, added explicit positions where necessary
* changed FunctionArgs param to accept any IDictionary instead of only a Hashtable
* disallowed the use of 'ENV' as a profile name
* updated CONTRIBUTING.md for new doc/help stuff
* added more EOL dates for version map
* upgraded to mkdocs-material 8.5.6 and removed old social card hack
* added native social card support
* readme and quickstart tweaks

+++ Posh-IBWAPI/Public/Remove-IBConfig.ps1
@@ -2,44 +2,64 @@ function Remove-IBConfig
 {
     [CmdletBinding()]
     param(
-        [Parameter(ParameterSetName='Specific',Position=0,ValueFromPipeline=$true,ValueFromPipelineByPropertyName=$true)]
+        [Parameter(ParameterSetName='Specific',Position=0,ValueFromPipeline,ValueFromPipelineByPropertyName)]
         [ValidateScript({Test-NonEmptyString $_ -ThrowOnFail})]
         [Alias('name')]
         [string]$ProfileName,
-        [Parameter(ParameterSetName='All',Mandatory=$true)]
+        [Parameter(ParameterSetName='All',Mandatory)]
         [switch]$AllProfiles
     )
 
+    Begin {
+        # Ignore these calls when running stateless with an environment variable
+        # based profile.
+        if ('ENV' -eq (Get-CurrentProfile)) {
+            Write-Warning "Remove-IBConfig is not available when using an environment variable based profile."
+        }
+    }
+
     Process {
+
+        # Ignore these calls when running stateless with an environment variable
+        # based profile.
+        if ('ENV' -eq (Get-CurrentProfile)) {
+            return
+        }
+
+        $profiles = Get-Profiles
 
         if ('All' -eq $PSCmdlet.ParameterSetName) {
 
-            if ($AllProfiles) {
-                Write-Verbose "Removing all connection profiles."
-
-                # delete the config file if it exists
-                $configFile = Get-ConfigFile
-                if (Test-Path $configFile) {
-                    Remove-Item $configFile -Force
-                }
-
-                Import-IBConfig
+            if (-not $AllProfiles) {
+                # For some reason they used -AllProfiles:$false which is
+                # weird but valid. So we'll just not do anything.
                 return
             }
 
-            # it's possible they called this with -AllProfiles:$false which is
-            # weird but valid and we'll just not do anything
+            Write-Verbose "Removing all connection profiles."
 
+            # remove the profiles from memory
+            foreach ($profName in @($profiles.Keys)) {
+                $profiles.Remove($profName)
+            }
+
+            # erase the current profile name
+            Set-CurrentProfile ([string]::Empty)
+
+            # delete the config file if it exists
+            $configFile = Get-ConfigFile
+            if (Test-Path $configFile) {
+                Remove-Item $configFile -Force
+            }
+
+            # persist the changes and return
+            Export-IBConfig
+            return
         }
 
         # decide which profile to remove
-        $profToRemove = Get-CurrentProfile
-        if ($ProfileName) {
-            $profToRemove = $ProfileName
-        }
-
-        Write-Verbose "Removing $profToRemove"
-        $profiles = Get-Profiles
+        $profToRemove = if ($ProfileName) { $ProfileName } else { Get-CurrentProfile }
+        Write-Verbose "Removing $profToRemove profile"
 
         if ($profToRemove -in $profiles.Keys) {
 
@@ -58,55 +78,13 @@ function Remove-IBConfig
             return
         }
 
-        # save changes to disk
-
         # if this is the last entry, just delete the config file
         $configFile = Get-ConfigFile
-        if ($profiles.Count -lt 1) {
-            if (Test-Path $configFile -PathType Leaf) {
-                Remove-Item $configFile -Force
-                Import-IBConfig
-            }
-        } else {
-            Export-IBConfig
+        if ($profiles.Count -eq 0 -and (Test-Path $configFile -PathType Leaf)) {
+            Remove-Item $configFile -Force
         }
 
+        # persist the changes
+        Export-IBConfig
     }
-
-
-
-    <#
-    .SYNOPSIS
-        Remove a WAPI connection profile.
-
-    .DESCRIPTION
-        When called with no parameters, the currently active connection profile will be removed.
-
-        When called with -ProfileName, the specified profile will be removed.
-
-        When called with -AllProfiles, all profiles will be removed.
-
-    .PARAMETER ProfileName
-        The name of the profile to remove.
-
-    .PARAMETER AllProfiles
-        If set, all profiles will be removed.
-
-    .EXAMPLE
-        Remove-IBConfig
-
-        Remove the currently active connection profile.
-
-    .EXAMPLE
-        Remove-IBConfig -AllHosts
-
-        Remove all connection profiles.
-
-    .LINK
-        Project: https://github.com/rmbolger/Posh-IBWAPI
-
-    .LINK
-        Set-IBConfig
-
-    #>
 }
